location improvements
Added ability to map a specific city, ST as an alternative to entering recruit ID. This way you can map a city location directly in cases where WIS location is not accurate or spelled incorrectly.

Also changed the Nominatim open street maps search to use "city" and "state" parameters in the query.

Also, when a location cannot be found, the user should see a properly error message displayed in the browser instead of generic "server error" message.