initial commit to implement async gathering of HTML pages
Redesigned the code to implement asyncio for grabbing HTML pages for game results. Instead of iterating through a list of game IDs and then URLs in a synchronous manner, this is now being done asynchronously usually in a manner for a few seconds for 10+ game IDs. Usually less than a second for just a handful of games. A few quick tests show that in production 10 games can take 60+ seconds to complete, but with this async code it can take less than 30 seconds.