changes to support game analysis selection toggles
new .js code to implement the switchbox toggles
the .js also identifies which games to submit, generates the URL with parameters to pass
the changes to the view now accept multiple gameids passed in as parameters and checks them against the known of games for that team. if the list of gameids passed aren't included in the teams overall list of games then the request fails.